Minor changes & fixes in SyncBinary; tweak `IScanner<T>`
I couldn't see any reason for `IScanner<T>.Read(_, 0, _)` to return a
buffer of length zero, so I changed the spec to say that it should
return at least one item if minLength is 0.

In SyncLib:
- Fix bugs in `SyncBinary.Writer`
- Bug fix: `SyncTimeSpanAsSeconds` and `SyncTimeSpanAsMinutes` wrote
  integers even if not requested
- Add `SyncManagerExt.Sync` overload for these parameters:
  `FieldId name, string? savable, ObjectMode listMode, int tupleLength`
  - Add `SyncLibStringBuilder` type used by this method
- Add tests in SyncBinaryWriterTests based on SyncBinary documentation
- Fix & clarify SyncBinary documentation

+++ Core/Loyc.Essentials/SyncLib/Binary/SyncBinary.Writer.cs
@@ -86,19 +86,19 @@ partial class SyncBinary
 
 		public BigInteger Sync(FieldId name, BigInteger savable) { _s.Write(savable); return savable; }
 
-		public char   Sync(FieldId name, char savable) { _s.Write((ushort)savable); return savable; }
+		public char   Sync(FieldId name, char savable) { _s.Write((uint)savable); return savable; }
 
 		public string? Sync(FieldId name, string? savable) { _s.Write(savable); return savable; }
 
 		public int Sync(FieldId name, int savable, int bits, bool signed = true)
 		{
-			_s.WriteBitfield(savable, (uint)bits);
+			_s.WriteBitfield((uint)savable, (uint)bits);
 			return savable;
 		}
 
 		public long Sync(FieldId name, long savable, int bits, bool signed = true)
 		{
-			_s.WriteBitfield(savable, (uint)bits);
+			_s.WriteBitfield((ulong)savable, (uint)bits);
 			return savable;
 		}
